// features/activities/types.ts
export type ActivityPreview = {
  id: string;
  name: string;
  location: string;
  description: string;
  price: string;
  currencySymbol: string;
  rating: number;
  reviewCount: number;
  duration: string;
  tags: string[];
};